kaasflow/frontend/requirements.txt
flask==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
bcrypt==4.1.2
pyjwt==2.8.0
requests==2.31.0
flask-cors==4.0.0
gunicorn==0.23.2